Move BlockGenerator and HashSetCasperTestNode to helpers
