+++ casper/src/test/scala/coop/rchain/casper/util/rholang/InterpreterUtilTest.scala
@@ -3,8 +3,8 @@ package coop.rchain.casper.util.rholang
 import com.google.protobuf.ByteString
 import InterpreterUtil._
 import coop.rchain.catscontrib.Capture._
-import coop.rchain.casper.{BlockDag, BlockGenerator}
-import coop.rchain.casper.BlockDagState._
+import coop.rchain.casper.BlockDag
+import coop.rchain.casper.helper.BlockDagState._
 import coop.rchain.casper.protocol._
 import coop.rchain.casper.util.ProtoUtil
 import coop.rchain.rholang.interpreter.Runtime
@@ -15,6 +15,7 @@ import cats.implicits._
 import cats.mtl.implicits._
 import java.nio.file.Files
 
+import coop.rchain.casper.helper.BlockGenerator
 import coop.rchain.casper.util.rholang.RuntimeManager.StateHash
 import monix.execution.Scheduler.Implicits.global
